apps/api/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt
pyjwt
pymongo

google-generativeai
google-api-core
google-auth
googleapis-common-protos
grpcio
grpcio-status
protobuf

aris-sdk
annotated-doc
stripe
pypdf
redis
email-validator
jinja2
annotated-types
requests
idna
certifi
urllib3
anyio
sniffio
h11
click
dnspython
httptools
uvloop
watchfiles
websockets
fsspec
semantic-version
tqdm
filelock
typing-extensions
markdown-it-py
mdurl
pygments
rich
svix
mcp
firecrawl-py
